requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==3.4.7
dj-rest-auth==4.0.0
Django==4.2
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
